Workaround for Windows's lack of %zu printf format specifier
git-svn-id: http://ludobloom.com/svn/StemLibProjects@1484 d291a34c-4c4d-0410-8556-a73de5c6a500

+++ resourcemanager/trunk/test_source/unittest/suites/ResourceManagerTest.c
@@ -159,11 +159,11 @@ static void testNULLResources() {
 	resourceToLoad = NULL;
 	
 	resourceManager->referenceResource(resourceManager, "type", resourceNameToLoad);
-	TestCase_assert(resourceManager->resourceCount == 0, "Expected 0 but got %zu\n", resourceManager->resourceCount);
+	TestCase_assert(resourceManager->resourceCount == 0, "Expected 0 but got %u\n", (unsigned int) resourceManager->resourceCount);
 	
 	resourceToLoad = "!NULL";
 	resourceManager->referenceResource(resourceManager, "type", resourceNameToLoad);
-	TestCase_assert(resourceManager->resourceCount == 1, "Expected 1 but got %zu\n", resourceManager->resourceCount);
+	TestCase_assert(resourceManager->resourceCount == 1, "Expected 1 but got %u\n", (unsigned int) resourceManager->resourceCount);
 	
 	resourceManager->dispose(resourceManager);
 }
